+++ src/app/sample/_components/Sol.tsx
@@ -2,7 +2,7 @@
 
 import { useRef } from 'react';
 import { Modal, ModalHandle } from '@/components/Modal';
-import { Notify } from '@/components/Toast/ToastProvider';
+import { notify } from '@/util/toast';
 
 export default function Sol() {
   const modalRef = useRef<ModalHandle | null>(null);
@@ -31,7 +31,7 @@ export default function Sol() {
         <div className="bg-white p-5">
           <button
             onClick={() => modalRef.current?.open()}
-            className="cursor-pointer rounded-md bg-gray-400 px-5 py-2 font-bold text-white"
+            className="cursor-pointer rounded-md bg-gray-600 px-5 py-2 font-semibold text-white"
           >
             모달 열기
           </button>
@@ -45,26 +45,45 @@ export default function Sol() {
           />
         </div>
       </div>
-      <div className="border border-gray-200 bg-gray-50 p-5">
+      <div className="mt-[15px] border border-gray-200 bg-gray-50 p-5">
         <h3 className="mb-5 text-xl font-bold">2. Toast</h3>
         <p className="mb-5 leading-7">
-          <b>
-            ✅ Props : type(타입:confirm,alert,content) | title(제목) | description(설명) | actionLabel(버튼명) |
-            onAction(버튼액션함수)
-          </b>
+          <b>✨ 기본형으로 통일하면 좋을 거 같아요. (but, 일단 type에 따라 아이콘이 나오게 설정해두었습니다😀)</b>
           <br />
-          1️⃣ 컨펌 모달 : type === confirm | title | description | actionLabel | onAction
+          <b>✅ 임포트 : {`import { notify } from '@/util/toast';`}</b>
           <br />
-          2️⃣ 확인 모달 : type === alert | title | description
+          1️⃣ 기본 토스트 : {`notify({ message: '기본 타입 알림입니다!' })`}
           <br />
-          3️⃣ 콘텐츠 모달 : type === content | children
+          2️⃣ 성공 토스트 : {`notify({ type: 'success', message: '기본 타입 알림입니다!' })`}
+          <br />
+          3️⃣ 에러 토스트 : {`notify({ type: 'error', message: '기본 타입 알림입니다!' })`}
+          <br />
+          4️⃣ 알림 토스트 : {`notify({ type: 'info', message: '기본 타입 알림입니다!' })`}
         </p>
         <div className="bg-white p-5">
           <button
-            onClick={() => Notify({ message: '기본 타입 알림입니다!' })}
-            className="cursor-pointer rounded-md bg-gray-400 px-5 py-2 font-bold text-white"
+            onClick={() => notify({ message: '기본 타입 알림입니다!' })}
+            className="cursor-pointer rounded-md bg-gray-800 px-5 py-2 font-semibold text-white"
           >
-            토스트 알림
+            기본 토스트
+          </button>
+          <button
+            onClick={() => notify({ type: 'success', message: '기본 타입 알림입니다!' })}
+            className="ml-[10px] cursor-pointer rounded-md bg-gray-800 px-5 py-2 font-semibold text-white"
+          >
+            성공 토스트
+          </button>
+          <button
+            onClick={() => notify({ type: 'error', message: '기본 타입 알림입니다!' })}
+            className="ml-[10px] cursor-pointer rounded-md bg-gray-800 px-5 py-2 font-semibold text-white"
+          >
+            에러 토스트
+          </button>
+          <button
+            onClick={() => notify({ type: 'info', message: '기본 타입 알림입니다!' })}
+            className="ml-[10px] cursor-pointer rounded-md bg-gray-800 px-5 py-2 font-semibold text-white"
+          >
+            알림 토스트
           </button>
         </div>
       </div>
